+++ packages/file_selector/file_selector/example/lib/open_image_page.dart
@@ -18,6 +18,7 @@ class OpenImagePage extends StatelessWidget {
     final XTypeGroup typeGroup = XTypeGroup(
       label: 'images',
       extensions: <String>['jpg', 'png'],
+      macUTIs: <String>['public.image'],
     );
     final XFile? file =
         await openFile(acceptedTypeGroups: <XTypeGroup>[typeGroup]);